Working on emulator enhacements
